feat: add search_name_override to series
This feature helps in situations where the commonly used show name for
torrents is different from the show name itself (or at least what info
providers tell us the show name is). While it'd be possible to just
change the show name, this lets us support keeping the show name while
having an override when searching.